Remove most methods from Stats and move to UserStats, leaving only abstract startCurrentSession(), endCurrentSession() and getCurrentSession().
