+++ src/main/java/dream/fcard/logic/stats/UserStats.java
@@ -14,7 +14,54 @@ public class UserStats extends Stats {
 
     /** Constructs a new instance of UserStats with no stored data. */
     public UserStats() {
-        super();
+        this.sessionList = new SessionList();
+        this.currentSession = null;
         System.out.println("New UserStats object created");
+    }
+
+    /** Sets the sessionList of the current Stats object to the given newSessionList. */
+    public void setSessionList(SessionList newSessionList) {
+        this.sessionList = newSessionList;
+    }
+
+    /** Gets the list of sessions. */
+    public SessionList getSessionList() {
+        return this.sessionList;
+    }
+
+    @Override
+    public void startCurrentSession() {
+        if (this.currentSession != null) {
+            endCurrentSession(); // should not occur, but should terminate just in case
+            System.out.println("Existing current session detected. Terminating it first...");
+        }
+
+        // debug (change to Logger when implemented)
+        System.out.println("Starting a session...");
+
+        this.currentSession = new Session(); // currentSession should be null
+    }
+
+    @Override
+    public void endCurrentSession() {
+        // assert current session is not null?
+        try {
+            this.currentSession.endSession();
+            this.sessionList.addSession(currentSession);
+
+            // reset currentSession to null since this is terminated
+            this.currentSession = null;
+
+            // debug (change to Logger when implemented)
+            System.out.println("Ending the current session...");
+        } catch (Exception e) {
+            e.printStackTrace();
+            System.out.println("Current session not found?");
+        }
+    }
+
+    @Override
+    public Session getCurrentSession() {
+        return this.currentSession;
     }
 }
